Include webxr layers polyfill.
Now the web version works immersively, but no controllers.

Also include WebXR main polyfill, in addition to layers polyfill (this
was an attempt to get controllers working, but still did not fix it; it
is unclear whether this polyfill is necessary, but I have seen advice
that it is better to include it).

I tried to include a local copy rather than linking to a CDN, but I
can't find a reliable way of ensuring js files are in the export
directory.