Major fix for ball bouncing randomly.
It turns out that balls were frequently bouncing off the 'shadow'
animatable body of the other ball, or in some cases even themselves,
because the animatable body was invisible but its collision sphere was
still active.

This would look like a ball randomly bouncing off to nowhere, but it can
be seen and reproduced clearly if debug collision shapes are turned on.

Disabled the ball objects when not in use.